refactored user and health tests to seperate file
