+++ lib/db/index.js
@@ -20,11 +20,11 @@ db.connect = async() => {
 
     db = mongoose.connection;
 
-    console.log( "DATABASE CONNECTED" );
+    console.log( "database connected" );
 
   } catch ( error ) {
 
-    console.error( "DATABASE CONNECTION FAILED: " + error.message );
+    console.error( "database connection failed: " + error.message );
   }
 };
