+++ lib/index.js
@@ -27,11 +27,11 @@ const start = async() => {
 
     app.listen( port );
 
-    console.log( "AGENDA API LISTENING ON PORT: " + port );
+    console.log( "agenda api listening on port: " + port );
 
   } catch ( error ) {
 
-    console.error( "AGENDA API STARTUP FAILED: " + error.message );
+    console.error( "agenda api failed to start: " + error.message );
   }
 };
