+++ test/tests/controllers/health.js
@@ -1,0 +1,14 @@
+const assert  = require( "assert" );
+const http    = require( "../../utils/http" );
+
+describe( "health.js", function() {
+
+  it( "should respond with ok", async() => {
+    const path = "/health";
+
+    const res = await http.get( path );
+
+    assert( res === "OK" );
+  });
+
+});
